Rename slice components
