+++ src/app/store.js
@@ -1,7 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
+import  currentArtReducer  from '../features/currentArt/currentArtSlice'
 
 export const store = configureStore({
-  reducer: {},
+  reducer: {
+    currentArt: currentArtReducer
+  },
 });
 
 export default store
